# Add performance test for internal & external flash for Read, write & Erase