@r-prabu 2024-06-14
> Hi @r-prabu , could you share whether this result is calculated with using SPI polling, or SPI DMA?

@edwakuwaku SPI DMA is not enabled in the above test.