@edwakuwaku 2024-06-14
With enabling the SPI DMA, the performance should rise, you may try and compare it.